prediction_pipeline/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
anthropic>=0.40.0
pyyaml>=6.0.1
lxml>=4.9.0
pytest>=7.4.0
